#23 add text drag
